Shorten errorCodes response property name and remove errorCodes column from db
